change student model, refactor api, add validator

+++ src/main/java/spring/demo/students/repository/StudentRepo.java
@@ -3,6 +3,6 @@ package spring.demo.students.repository;
 import spring.demo.students.model.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepo extends JpaRepository<Student, Integer> {
+public interface StudentRepo extends JpaRepository<Student, Long> {
 
 }
